sql injections fix

+++ server/phpWebsite/delete_alarm.php
@@ -1,10 +1,11 @@
 <?php
-include("conn.php");
+include ("conn.php");
 session_start();
 $value = $_POST["value"];
-$q = "DELETE FROM Alarms WHERE id = $value";
-$r = mysqli_query($conn,$q);
 
-
+if (is_numeric($value)) {
+    $q = "DELETE FROM Alarms WHERE id = $value";
+    $r = mysqli_query($conn, $q);
+}
 
 ?>
