+++ server/phpWebsite/delete_contact.php
@@ -1,9 +1,12 @@
 <?php
-include("conn.php");
+include ("conn.php");
 session_start();
 $value = $_POST["value"];
-$q = "DELETE FROM Contacts WHERE id = $value";
-$r = mysqli_query($conn,$q);
+if (is_numeric($value)) {
+    $q = "DELETE FROM Contacts WHERE id = $value";
+    $r = mysqli_query($conn, $q);
+
+}
