+++ server/phpWebsite/delete_pattern.php
@@ -1,10 +1,10 @@
 <?php
 
-include("conn.php");
+include ("conn.php");
 
 $value = $_POST["value"];
-$q = "DELETE FROM patterns_intent WHERE id = $value";
-$r = mysqli_query($conn,$q);
-
+if (is_numeric($value)) {
+    $q = "DELETE FROM patterns_intent WHERE id = $value";
+    $r = mysqli_query($conn, $q);
+}
 ?>
-
